DynamicParameterMatcher stub, frontend
