+++ src/frontend/DynamicParameterMatcher.cpp
@@ -1,0 +1,12 @@
+//
+// Created by Piotr Kubala on 07/01/2023.
+//
+
+#include "DynamicParameterMatcher.h"
+
+using namespace pyon::matcher;
+
+
+pyon::matcher::MatcherAlternative DynamicParameterMatcher::create() {
+    return MatcherAlternative();
+}
